fix export bug

+++ util.py
@@ -13,4 +13,4 @@ def display_audio(samples: tp.List[torch.Tensor], path: str = None):
     ipd.display(ipd.Audio(audio, rate=32000))
 
     if path:
-        write(path, 32000, audio)
+        write(path, 32000, audio.astype(np.float32))
